Add basic WPT coverage for b2 web bundles.
We're adding some basic WPTs for b2 bundles before we deprecate the "b1"
format completely. Follow-up CLs will also handle other directories with
web bundle tests, if needed.
This is a step 1 of the current plan for e2e test coverage for 'b2'
bundles. The whole plan is in the attached bug.

Bug: 1249349
Change-Id: I8ea4fd78aabde06bb22a8267b8a1ae4300298f5e